Move srs_string_dumps_hex from app to protocol utility.
